src/main/java/nextstep/courses/domain/Course.java
package nextstep.courses.domain;

import nextstep.payments.domain.Payments;
import nextstep.sessions.domain.Session;
import nextstep.sessions.domain.Sessions;

import java.time.LocalDateTime;

public class Course {
    private Long id;

    private String title;

    private Long creatorId;

    private LocalDateTime createdAt;

    private LocalDateTime updatedAt;

    private Sessions sessions;

    public Course() {
    }

    public Course(String title, Long creatorId) {
        this(0L, title, creatorId, LocalDateTime.now(), null);
    }

    public Course(Long id, String title, Long creatorId, LocalDateTime createdAt, LocalDateTime updatedAt) {
        this.id = id;
        this.title = title;
        this.creatorId = creatorId;
        this.createdAt = createdAt;
        this.updatedAt = updatedAt;
        this.sessions = new Sessions();
    }

    public Course(String title, Long creatorId, Sessions sessions) {
        this.id = 0L;
        this.title = title;
        this.creatorId = creatorId;
        this.createdAt = LocalDateTime.now();
        this.updatedAt = null;
        this.sessions = sessions;
    }

    public String getTitle() {
        return title;
    }

    public Long getCreatorId() {
        return creatorId;
    }

    public LocalDateTime getCreatedAt() {
        return createdAt;
    }

    public void addSession(Session addSession) {
        sessions.addSession(addSession);
    }

    public Payments pay() {
        return sessions.pay(creatorId);
    }

    @Override
    public String toString() {
        return "Course{" +
                "id=" + id +
                ", title='" + title + '\'' +
                ", creatorId=" + creatorId +
                ", createdAt=" + createdAt +
                ", updatedAt=" + updatedAt +
                '}';
    }
}
